perf(dashboard): lazy-load markdown renderer
